chore: 清理死代码 + agent 身份 block 构造器归位 agent-assembly
- 删除死代码：buildSkillCatalog/SkillItem（skill capability 自渲染后残留）、serializeSystemPrompt（无人引用的转发壳）、kernel 零引用导出（taggedBlock/pipeMiddleware/lastEntry/McpHost）
- context/blocks.ts 归位：project-instructions/agent-profile/session-context/preloaded-context 是 assembly 私有的 agent 身份 block（非 capability 贡献面），并入 session/agent-assembly.ts；ContextFile 类型归 read-context-files（磁盘读取结果，属 context 层）
- 导出面收缩后 core 820 tests / verify 全绿
