+++ packages/core/src/kernel/context-block.ts
@@ -3,15 +3,6 @@ export interface ContextBlock {
   render(): string;
 }
 
-export function taggedBlock(kind: string, content: string): ContextBlock {
-  return {
-    kind,
-    render() {
-      return `<${kind}>\n${content}\n</${kind}>`;
-    },
-  };
-}
-
 export function serializeBlocks(blocks: ReadonlyArray<ContextBlock | null>): string {
   const rendered = blocks
     .filter((b): b is ContextBlock => b !== null)
